08-assembler/pyproject.toml
[project]
name = "08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyparsing>=3.2.3",
    "pyyaml>=6.0.2",
    "ruamel-yaml>=0.18.14",
]
